FreePBX Fax Upload to SFTP
This one was requested by a customer. They wanted their faxes to be uploaded to a communal SFTP server because they did not like the email or web access offered by FreePBX. 

The script monitors the fax directory on a FreePBX server. It ignores the tiff versions of the faxes (customer request), and uploads the PDF versions to a SFTP server. There are checks to make sure the upload does not happen while the PDF file is still generating, plus a file size check after the upload is finished to make sure it matches the original.